refactor: reorganize data/ file structure
