+++ src/data/enemy.js
@@ -1,0 +1,144 @@
+/* Enemy Data */
+
+
+export const ENEMY2_SPRITES = {
+	STAND_CYCLE: [
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,2,2,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,3,2,3,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,2,2,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,
+			    0,0,0,1,0,0,0,0,2,0,0,0,0,1,0,0,
+			    0,0,0,0,1,1,2,2,2,2,2,1,1,0,0,0,
+			    0,0,0,0,0,0,1,2,2,2,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,2,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,1,2,2,2,1,0,0,0,0,0,
+			    0,0,0,0,0,2,0,0,1,0,0,2,0,0,0,0,
+			    0,0,0,0,0,0,0,1,2,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,2,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,2,2,0,0,0,0,0,0,
+			    0,0,0,0,0,0,1,1,1,1,1,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			frameDuration: 80,
+		},
+	]
+};
+
+
+export const SLOTH_SPRITES = {
+	STAND_CYCLE: [
+		{
+		    sprite: [
+		        0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+		        0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+		        0,0,0,0,0,0,0,0,0,0,5,5,0,0,0,0,
+		        0,0,0,0,0,0,0,0,0,5,0,0,6,0,0,0,
+		        0,0,0,0,0,0,4,5,4,5,5,0,0,0,0,0,
+		        0,0,0,0,0,5,5,4,5,4,3,3,3,0,0,0,
+		        0,0,0,0,4,5,3,5,3,5,3,1,3,1,0,0,
+		        0,0,0,5,5,4,3,5,3,5,3,3,0,3,0,0,
+		        0,0,5,4,3,5,3,4,3,4,3,0,0,0,3,0,
+		        0,4,3,5,2,4,3,0,3,0,0,0,0,0,0,0,
+		        0,5,2,5,4,0,3,0,0,3,3,0,0,0,0,0,
+		        0,3,4,4,0,0,3,0,0,0,3,0,0,0,0,0,
+		        0,3,0,3,0,0,3,0,0,0,3,0,0,0,0,0,
+		        0,0,3,3,0,3,3,0,0,0,0,0,0,0,0,0,
+		        0,0,3,0,0,3,0,0,0,0,0,0,0,0,0,0,
+		        0,0,3,3,0,3,3,0,0,0,0,0,0,0,0,0,
+		    ],
+		    width: 16,
+		    frameDuration: 15,
+		    height: 16,
+		    tPad: 2,
+		    bPad: 0,
+		    lPad: 1,
+		    rPad: 1,
+		},
+		{
+		    sprite: [
+		        0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+		        0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+		        0,0,0,0,0,0,0,0,0,0,5,5,0,0,0,0,
+		        0,0,0,0,0,0,0,0,0,5,0,0,6,0,0,0,
+		        0,0,0,0,0,5,5,5,5,5,4,0,0,0,0,0,
+		        0,0,0,5,5,5,5,5,5,4,3,3,3,0,0,0,
+		        0,0,0,5,4,5,3,5,3,5,3,1,3,1,0,0,
+		        0,0,5,5,5,4,3,5,3,5,3,3,0,3,0,0,
+		        0,0,5,4,3,5,3,4,3,4,3,0,3,0,3,0,
+		        0,4,5,5,2,4,3,0,3,0,0,0,0,0,3,0,
+		        0,5,2,5,4,0,3,0,3,3,0,0,0,3,0,0,
+		        0,3,4,4,0,0,3,0,0,3,0,0,0,0,0,0,
+		        0,3,0,3,0,0,3,0,0,3,0,0,0,0,0,0,
+		        0,0,3,3,0,3,3,0,0,0,0,0,0,0,0,0,
+		        0,0,3,0,0,3,0,0,0,0,0,0,0,0,0,0,
+		        0,0,3,3,0,3,3,0,0,0,0,0,0,0,0,0,
+		    ],
+		    width: 16,
+		    frameDuration: 14,
+		    height: 16,
+		    tPad: 2,
+		    bPad: 0,
+		    lPad: 1,
+		    rPad: 1,
+		},
+		{
+		    sprite: [
+		        0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+		        0,0,0,0,0,0,0,0,0,0,0,5,0,0,0,0,
+		        0,0,0,0,0,0,0,5,5,0,5,0,6,0,0,0,
+		        0,0,0,0,0,5,5,5,5,5,0,0,0,0,0,0,
+		        0,0,0,0,5,5,4,5,4,5,5,0,0,0,0,0,
+		        0,0,0,0,5,5,5,4,5,4,3,3,3,0,0,0,
+		        0,0,0,5,4,5,3,4,3,5,3,1,3,1,0,0,
+		        0,0,0,5,5,4,3,4,3,5,3,3,0,3,0,0,
+		        0,0,5,4,3,4,3,4,3,4,3,3,3,3,0,0,
+		        0,0,3,5,2,4,3,0,3,0,0,0,0,0,3,0,
+		        0,0,2,5,4,0,3,0,3,0,0,0,0,0,3,0,
+		        0,3,4,4,0,0,3,0,3,3,0,0,0,3,3,0,
+		        0,0,0,3,0,0,3,0,0,3,3,0,0,0,0,0,
+		        0,0,3,3,0,0,3,0,0,0,0,0,0,0,0,0,
+		        0,0,3,0,0,3,0,0,0,0,0,0,0,0,0,0,
+		        0,0,3,0,0,3,3,0,0,0,0,0,0,0,0,0,
+		    ],
+		    width: 16,
+		    frameDuration: 20,
+		    height: 16,
+		    tPad: 1,
+		    bPad: 0,
+		    lPad: 1,
+		    rPad: 1,
+		},
+		{
+		    sprite: [
+		        0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+		        0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+		        0,0,0,0,0,0,0,0,0,0,5,5,0,0,0,0,
+		        0,0,0,0,0,0,0,0,0,5,0,0,6,0,0,0,
+		        0,0,0,0,0,5,5,5,5,5,4,0,0,0,0,0,
+		        0,0,0,5,5,5,5,5,5,4,3,3,3,0,0,0,
+		        0,0,0,5,4,5,3,5,3,5,3,1,3,1,0,0,
+		        0,0,5,5,5,4,3,5,3,5,3,3,0,3,0,0,
+		        0,0,5,4,3,5,3,4,3,4,3,0,3,0,3,0,
+		        0,4,5,5,2,4,3,0,3,0,0,0,0,0,3,0,
+		        0,5,2,5,4,0,3,0,3,3,0,0,0,3,0,0,
+		        0,3,4,4,0,0,3,0,0,3,0,0,0,0,0,0,
+		        0,3,0,3,0,0,3,0,0,3,0,0,0,0,0,0,
+		        0,0,3,3,0,3,3,0,0,0,0,0,0,0,0,0,
+		        0,0,3,0,0,3,0,0,0,0,0,0,0,0,0,0,
+		        0,0,3,3,0,3,3,0,0,0,0,0,0,0,0,0,
+		    ],
+		    width: 16,
+		    frameDuration: 14,
+		    height: 16,
+		    tPad: 2,
+		    bPad: 0,
+		    lPad: 1,
+		    rPad: 1,
+		},
+	],
+}
