+++ src/data/player.js
@@ -1,0 +1,1407 @@
+/* Player Data */
+
+const RANGER_STAND_CYCLE = [60,60,60,60];
+const RANGER_SLIDE_CYCLE = [90000];
+const RANGER_JUMP_CYCLE = [12, 7, 10, 7, 7, 5, 90000];
+
+const RANGER_CROUCH_CYCLE = [2, 3, 3, 3, 3, 3, 90000];
+const RANGER_TURN_CYCLE = [8];
+const RANGER_WALK_FRAMES = [6, 7, 8];
+const RANGER_WALK_SHOOT_FRAMES = RANGER_WALK_FRAMES;
+const RANGER_SHOOT_FRAMES = [4, 4, 4, 4, 12, 15];
+export const RANGER_SPRITES = {
+	STAND_CYCLE: [
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,
+			    0,0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,5,4,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,5,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,1,1,1,1,1,1,1,0,0,
+			    0,0,0,0,0,0,5,5,1,1,1,5,1,0,0,0,
+			    0,0,0,0,0,0,0,1,6,3,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,6,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,1,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,0,2,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 1,
+			bPad: 0,
+			lPad: 5,
+			rPad: 2,
+			frameDuration: RANGER_STAND_CYCLE[0],
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,
+			    0,0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,5,4,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,5,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,1,1,1,1,1,1,1,0,0,
+			    0,0,0,0,0,0,5,5,1,1,1,5,1,0,0,0,
+			    0,0,0,0,0,0,0,1,6,3,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,6,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,6,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,0,2,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 0,
+			bPad: 0,
+			lPad: 5,
+			rPad: 2,
+			frameDuration: RANGER_STAND_CYCLE[2],
+		},
+	],
+	CROUCH_CYCLE: [
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,
+			    0,0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,5,4,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,5,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,1,1,1,1,1,1,1,0,0,
+			    0,0,0,0,0,0,5,5,1,1,1,5,1,0,0,0,
+			    0,0,0,0,0,0,0,1,6,3,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,6,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,1,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,0,2,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 1,
+			bPad: 0,
+			lPad: 5,
+			rPad: 2,
+			frameDuration: RANGER_CROUCH_CYCLE[0],
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,
+			    0,0,0,0,0,0,0,5,5,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,5,4,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,1,1,1,1,1,1,1,0,0,
+			    0,0,0,0,0,0,5,5,1,1,1,5,1,0,0,0,
+			    0,0,0,0,0,0,0,1,6,3,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,6,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,3,6,3,3,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,0,2,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 1,
+			bPad: 0,
+			lPad: 5,
+			rPad: 2,
+			frameDuration: RANGER_CROUCH_CYCLE[1],
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,
+			    0,0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,5,4,1,1,1,1,1,0,0,
+			    0,0,0,0,0,0,3,1,1,1,1,5,1,0,0,0,
+			    0,0,0,0,0,0,5,5,1,3,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,6,3,0,0,0,0,0,0,
+			    0,0,0,0,0,3,3,6,6,6,3,3,0,0,0,0,
+			    0,0,0,0,0,0,0,6,6,6,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,0,2,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 3,
+			bPad: 0,
+			lPad: 5,
+			rPad: 2,
+			frameDuration: RANGER_CROUCH_CYCLE[2],
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,
+			    0,0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,1,1,1,1,1,1,1,0,0,
+			    0,0,0,0,3,0,5,5,1,1,1,5,1,0,0,0,
+			    0,0,0,0,0,3,3,1,6,3,1,3,0,0,0,0,
+			    0,0,0,0,0,0,3,3,6,3,3,0,0,0,0,0,
+			    0,0,0,0,0,0,0,6,6,6,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,0,2,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 4,
+			bPad: 0,
+			lPad: 4,
+			rPad: 2,
+			frameDuration: RANGER_CROUCH_CYCLE[3],
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,
+			    0,0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,5,4,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,1,1,1,1,1,1,1,0,0,
+			    0,0,0,0,0,0,5,5,1,1,1,5,1,0,0,0,
+			    0,0,0,0,0,3,3,1,6,3,1,3,0,0,0,0,
+			    0,0,0,0,0,0,0,6,6,6,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,0,2,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 4,
+			bPad: 0,
+			lPad: 5,
+			rPad: 2,
+			frameDuration: RANGER_CROUCH_CYCLE[4],
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,
+			    0,0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,5,4,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,1,1,1,1,1,1,1,0,0,
+			    0,0,0,0,0,0,5,5,1,1,1,5,1,0,0,0,
+			    0,0,0,0,0,0,0,1,6,3,1,0,0,0,0,0,
+			    0,0,0,0,0,0,3,6,6,6,3,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,0,2,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 4,
+			bPad: 0,
+			lPad: 5,
+			rPad: 2,
+			frameDuration: RANGER_CROUCH_CYCLE[5],
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,
+			    0,0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,5,4,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,5,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,1,1,1,1,1,1,1,0,0,
+			    0,0,0,0,0,0,5,5,1,1,1,5,1,0,0,0,
+			    0,0,0,0,0,0,0,1,6,3,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,6,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,0,2,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 3,
+			bPad: 0,
+			lPad: 5,
+			rPad: 2,
+			frameDuration: RANGER_CROUCH_CYCLE[6],
+		},
+	],
+	JUMP_CYCLE: [
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,
+			    0,0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,5,4,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,5,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,1,1,1,1,1,1,1,0,0,
+			    0,0,0,0,0,0,5,5,1,1,1,5,1,0,0,0,
+			    0,0,0,0,0,0,0,1,6,3,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,1,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,0,2,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 3,
+			bPad: 0,
+			lPad: 5,
+			rPad: 2,
+			frameDuration: RANGER_JUMP_CYCLE[0],
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,
+			    0,0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,5,4,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,1,1,1,1,1,1,1,0,0,
+			    0,0,0,0,0,0,5,5,1,1,1,5,1,0,0,0,
+			    0,0,0,0,0,0,0,1,6,3,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,1,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,0,2,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 2,
+			bPad: 0,
+			lPad: 5,
+			rPad: 2,
+			frameDuration: RANGER_JUMP_CYCLE[1],
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,
+			    0,0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,1,1,1,1,1,1,1,0,0,
+			    0,0,0,0,0,0,5,5,1,1,1,5,1,0,0,0,
+			    0,0,0,0,0,0,0,1,6,3,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,6,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,1,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,2,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 1,
+			bPad: 1,
+			lPad: 5,
+			rPad: 2,
+			frameDuration: RANGER_JUMP_CYCLE[2],
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,
+			    0,0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,5,4,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,1,1,1,1,1,1,1,0,0,
+			    0,0,0,0,0,0,5,5,1,1,1,5,1,0,0,0,
+			    0,0,0,0,0,0,0,1,6,3,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,6,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,1,1,1,1,0,0,0,0,0,
+			    0,0,0,0,0,3,3,1,1,1,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,2,0,0,0,0,0,0,
+			    0,0,0,0,0,0,2,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 0,
+			bPad: 3,
+			lPad: 5,
+			rPad: 2,
+			frameDuration: RANGER_JUMP_CYCLE[3],
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,
+			    0,0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,5,4,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,1,1,1,1,1,1,1,0,0,
+			    0,0,0,0,0,0,5,5,1,1,1,5,1,0,0,0,
+			    0,0,0,0,0,0,0,1,6,3,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,6,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,3,6,3,0,0,0,0,0,0,
+			    0,0,0,0,0,3,3,1,1,1,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,2,0,0,0,0,0,0,
+			    0,0,0,0,0,0,2,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 0,
+			bPad: 2,
+			lPad: 5,
+			rPad: 2,
+			frameDuration: RANGER_JUMP_CYCLE[4],
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,
+			    0,0,0,0,0,0,0,5,5,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,1,1,1,1,1,1,1,0,0,
+			    0,0,0,0,0,0,5,5,1,1,1,5,1,0,0,0,
+			    0,0,0,0,0,0,3,1,6,3,1,0,0,0,0,0,
+			    0,0,0,0,0,3,3,3,6,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,3,6,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,6,6,6,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,2,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 0,
+			bPad: 1,
+			lPad: 5,
+			rPad: 2,
+			frameDuration: RANGER_JUMP_CYCLE[5],
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,
+			    0,0,0,0,0,0,0,5,5,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,5,4,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,1,1,1,1,1,1,1,0,0,
+			    0,0,0,0,0,0,5,5,1,1,1,5,1,0,0,0,
+			    0,0,0,0,0,3,3,1,6,3,1,0,0,0,0,0,
+			    0,0,0,0,0,0,3,3,6,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,1,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,0,2,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 1,
+			bPad: 0,
+			lPad: 5,
+			rPad: 2,
+			frameDuration: RANGER_JUMP_CYCLE[6],
+		},
+	],
+	TURN_CYCLE: [
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,2,2,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,
+			    0,0,0,0,0,0,5,1,4,1,5,0,0,0,0,0,
+			    0,0,0,0,0,0,0,5,4,5,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,5,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,3,1,3,3,0,0,0,0,0,
+			    0,0,0,0,0,0,5,3,5,5,5,0,0,0,0,0,
+			    0,0,0,0,0,0,0,5,1,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,6,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,1,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,0,2,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 0,
+			bPad: 0,
+			lPad: 5,
+			rPad: 4,
+			frameDuration: RANGER_TURN_CYCLE[0],
+		},
+	],
+	SLIDE_CYCLE: [
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,0,
+			    0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,0,
+			    0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,5,4,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,3,1,1,1,1,1,1,1,0,0,0,
+			    0,0,0,0,0,5,5,1,1,1,5,1,0,0,0,0,
+			    0,0,0,0,0,0,1,6,3,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,6,3,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,1,3,3,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,1,0,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,0,0,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,2,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 2,
+			bPad: 0,
+			lPad: 4,
+			rPad: 3,
+			frameDuration: RANGER_SLIDE_CYCLE[0],
+		},
+	],
+	WALK_CYCLE: [
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,
+			    0,0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,5,4,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,5,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,1,1,1,1,1,1,1,0,0,
+			    0,0,0,0,0,0,5,5,1,1,1,5,1,0,0,0,
+			    0,0,0,0,0,0,0,1,6,3,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,6,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,1,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,2,0,0,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,2,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 1,
+			bPad: 0,
+			lPad: 5,
+			rPad: 2,
+			frameDuration: RANGER_WALK_FRAMES[0],
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,
+			    0,0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,5,4,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,1,1,1,1,1,1,1,0,0,
+			    0,0,0,0,0,0,5,5,1,1,1,5,1,0,0,0,
+			    0,0,0,0,0,0,0,1,6,3,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,6,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,6,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,2,0,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 1,
+			bPad: 0,
+			lPad: 5,
+			rPad: 2,
+			frameDuration: RANGER_WALK_FRAMES[1],
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,
+			    0,0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,5,4,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,5,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,1,1,1,1,1,1,1,0,0,
+			    0,0,0,0,0,0,5,5,1,1,1,5,1,0,0,0,
+			    0,0,0,0,0,0,0,1,6,3,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,6,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,1,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,2,0,0,0,0,0,0,
+			    0,0,0,0,0,0,2,0,0,0,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 0,
+			bPad: 0,
+			lPad: 5,
+			rPad: 2,
+			frameDuration: RANGER_WALK_FRAMES[2],
+		},
+	],
+	SHOOT_CYCLE: [
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,
+			    0,0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,5,4,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,5,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,1,1,1,1,1,1,1,0,0,
+			    0,0,0,0,0,0,5,5,1,1,1,5,1,0,0,0,
+			    0,0,0,0,0,0,0,1,6,3,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,6,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,1,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,
+			    0,0,0,0,0,0,2,0,0,0,2,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 1,
+			bPad: 0,
+			lPad: 5,
+			rPad: 2,
+			frameDuration: RANGER_SHOOT_FRAMES[0],
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,
+			    0,0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,5,4,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,5,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,1,1,1,1,1,1,1,7,0,
+			    0,0,0,0,0,0,5,5,1,1,1,5,1,0,0,0,
+			    0,0,0,0,0,0,0,1,6,3,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,6,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,1,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,
+			    0,0,0,0,0,0,2,0,0,0,2,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 1,
+			bPad: 0,
+			lPad: 5,
+			rPad: 1,
+			frameDuration: RANGER_SHOOT_FRAMES[1],
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,
+			    0,0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,5,4,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,5,3,0,0,0,0,8,7,
+			    0,0,0,0,0,3,1,1,1,1,1,1,1,8,7,8,
+			    0,0,0,0,0,5,5,1,1,1,5,1,0,0,0,7,
+			    0,0,0,0,0,0,1,3,6,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,6,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,1,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,
+			    0,0,0,0,0,0,2,0,0,0,2,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 1,
+			bPad: 0,
+			lPad: 5,
+			rPad: 0,
+			frameDuration: RANGER_SHOOT_FRAMES[2],
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,2,2,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,1,1,1,0,0,0,0,0,0,0,0,
+			    0,0,0,2,2,2,2,2,2,2,0,0,0,0,0,0,
+			    0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,5,4,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,3,3,5,3,0,0,0,0,0,0,0,
+			    0,0,0,0,0,3,3,6,1,1,1,1,1,0,0,0,
+			    0,0,0,0,0,3,1,1,1,1,5,1,0,0,0,0,
+			    0,0,0,0,0,5,5,1,6,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,1,3,6,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,1,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,
+			    0,0,0,0,0,0,2,0,0,0,2,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 1,
+			bPad: 0,
+			lPad: 3,
+			rPad: 3,
+			frameDuration: RANGER_SHOOT_FRAMES[3],
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,0,
+			    0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,0,
+			    0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,5,4,0,0,1,1,0,0,0,0,
+			    0,0,0,0,0,0,3,5,1,1,1,0,0,0,0,0,
+			    0,0,0,0,0,3,3,1,1,1,5,0,0,0,0,0,
+			    0,0,0,0,0,5,5,5,3,0,1,0,0,0,0,0,
+			    0,0,0,0,0,0,3,1,3,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,6,3,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,1,3,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,
+			    0,0,0,0,0,0,2,0,0,0,2,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 1,
+			bPad: 0,
+			lPad: 4,
+			rPad: 4,
+			frameDuration: RANGER_SHOOT_FRAMES[4],
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,
+			    0,0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,5,4,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,5,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,3,6,1,1,1,1,1,0,0,
+			    0,0,0,0,0,0,3,1,1,1,1,5,1,0,0,0,
+			    0,0,0,0,0,0,5,5,1,3,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,6,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,1,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,
+			    0,0,0,0,0,0,2,0,0,0,2,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 1,
+			bPad: 0,
+			lPad: 5,
+			rPad: 2,
+			frameDuration: RANGER_SHOOT_FRAMES[5],
+		},
+	],
+	WALK_SHOOT_CYCLE: [
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,
+			    0,0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,5,4,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,5,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,1,1,1,1,1,1,1,7,0,
+			    0,0,0,0,0,0,5,5,1,1,1,5,1,0,0,0,
+			    0,0,0,0,0,0,0,1,6,3,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,6,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,1,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,2,0,0,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,2,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 1,
+			bPad: 0,
+			lPad: 5,
+			rPad: 1,
+			frameDuration: RANGER_WALK_SHOOT_FRAMES[0],
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,
+			    0,0,0,0,0,0,0,5,5,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,5,4,0,0,0,0,0,8,7,
+			    0,0,0,0,0,3,1,1,1,1,1,1,1,8,7,8,
+			    0,0,0,0,0,5,5,1,1,1,5,1,0,0,0,7,
+			    0,0,0,0,0,0,1,3,6,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,6,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,6,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,2,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,2,0,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 0,
+			bPad: 0,
+			lPad: 5,
+			rPad: 0,
+			frameDuration: RANGER_WALK_SHOOT_FRAMES[1],
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,2,2,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,0,
+			    0,0,0,0,2,2,2,2,2,2,2,0,0,0,0,0,
+			    0,0,0,0,0,5,1,4,1,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,5,4,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,3,3,5,1,1,1,1,1,0,0,0,
+			    0,0,0,0,0,3,1,1,1,1,5,1,0,0,0,0,
+			    0,0,0,0,0,5,5,1,3,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,1,6,3,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,
+			    0,0,0,0,0,0,0,1,0,2,0,0,0,0,0,0,
+			    0,0,0,0,0,0,2,0,0,0,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 1,
+			bPad: 0,
+			lPad: 4,
+			rPad: 3,
+			frameDuration: RANGER_WALK_SHOOT_FRAMES[2],
+		},
+	],
+};
+
+const METROID_JUMP_FRAMES = [5, 5, 5, 5];
+const METROID_RUN_FRAMES = [6, 6, 6];
+export const METROID_SPRITES = {
+	STAND_CYCLE: [
+		{
+			sprite: [
+			    0,0,0,0,0,0,4,3,3,3,0,0,0,0,0,0,
+			    0,0,0,0,0,4,3,3,3,2,4,0,0,0,0,0,
+			    0,0,0,0,0,3,3,3,3,1,2,0,0,0,0,0,
+			    0,0,0,0,0,0,3,4,3,3,3,0,0,0,0,0,
+			    0,0,0,0,0,4,4,4,4,3,0,4,2,2,4,2,
+			    0,0,0,0,0,4,4,4,4,4,4,4,2,4,2,2,
+			    0,0,0,0,0,0,4,4,4,4,0,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,3,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,4,4,4,0,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,0,4,0,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,0,4,4,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,0,3,4,4,0,0,0,0,
+			    0,0,0,0,0,4,4,0,0,0,4,0,0,0,0,0,
+			    0,0,0,3,3,4,0,0,0,4,4,0,0,0,0,0,
+			    0,0,0,3,3,0,0,0,0,3,3,0,0,0,0,0,
+			    0,0,0,0,3,3,0,0,0,0,3,3,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			frameDuration: 20,
+			tPad: 0,
+			bPad: 0,
+			lPad: 3,
+			rPad: 0,
+		},
+	],
+	WALK_CYCLE: [
+		{
+			sprite: [
+			    0,0,0,0,0,0,4,3,3,3,0,0,0,0,0,0,
+			    0,0,0,0,0,4,3,3,3,2,4,0,0,0,0,0,
+			    0,0,0,0,0,3,3,3,3,1,2,0,0,0,0,0,
+			    0,0,0,0,0,0,3,2,3,3,3,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,4,3,0,0,0,0,0,0,
+			    0,0,0,0,0,3,3,3,4,4,2,0,0,0,0,0,
+			    0,0,0,0,0,3,4,4,4,2,2,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,3,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,4,4,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,4,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,4,0,4,3,0,0,0,0,0,0,
+			    0,0,0,0,0,3,0,4,4,0,0,0,0,0,0,0,
+			    0,0,0,0,3,2,4,4,0,0,0,0,0,0,0,0,
+			    0,0,0,0,3,3,4,3,0,0,0,0,0,0,0,0,
+			    0,0,0,0,3,4,3,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,3,3,3,0,0,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			frameDuration: METROID_RUN_FRAMES[0],
+			tPad: 0,
+			bPad: 0,
+			lPad: 4,
+			rPad: 5,
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,4,3,3,3,0,0,0,0,0,0,
+			    0,0,0,0,0,4,3,3,3,2,4,0,0,0,0,0,
+			    0,0,0,0,0,3,3,3,3,1,2,0,0,0,0,0,
+			    0,0,0,0,0,2,3,3,3,3,3,0,0,0,0,0,
+			    0,0,0,0,4,4,4,4,4,3,0,0,0,0,0,0,
+			    0,0,0,0,4,4,4,3,4,4,0,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,4,4,3,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,3,4,3,0,0,0,0,0,
+			    0,0,0,0,0,4,4,4,4,3,4,4,0,0,0,0,
+			    0,0,0,0,0,4,4,4,4,0,4,4,0,0,0,0,
+			    0,0,0,0,0,0,4,4,0,4,4,0,0,0,0,0,
+			    0,0,0,0,0,4,4,0,3,3,0,0,0,0,0,0,
+			    0,0,0,3,4,4,0,0,3,3,3,0,0,0,0,0,
+			    0,0,3,3,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,3,3,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,3,3,0,0,0,0,0,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			frameDuration: METROID_RUN_FRAMES[1],
+			tPad: 0,
+			bPad: 0,
+			lPad: 2,
+			rPad: 4,
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,4,3,3,3,0,0,0,0,0,
+			    0,0,0,0,0,0,4,3,3,3,2,4,0,0,0,0,
+			    0,0,0,0,0,0,3,3,3,3,1,2,0,0,0,0,
+			    0,0,0,0,0,0,2,1,3,3,3,3,0,0,0,0,
+			    0,0,0,0,4,4,4,4,1,2,3,0,0,0,0,0,
+			    0,0,0,3,3,4,2,3,2,4,0,0,0,0,0,0,
+			    0,0,0,0,0,0,2,2,2,0,0,0,0,0,0,0,
+			    0,0,0,0,0,4,4,4,3,0,0,0,0,0,0,0,
+			    0,0,0,0,0,4,4,4,4,0,0,0,0,0,0,0,
+			    0,0,3,3,0,0,4,4,4,4,0,0,0,0,0,0,
+			    0,0,3,4,0,0,4,0,0,4,0,0,0,0,0,0,
+			    0,3,3,4,4,4,0,0,4,4,3,0,0,0,0,0,
+			    0,3,0,0,4,4,0,0,4,4,3,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,3,4,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,3,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,3,3,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			frameDuration: METROID_RUN_FRAMES[2],
+			tPad: 0,
+			bPad: 0,
+			lPad: 1,
+			rPad: 4,
+		},
+	],
+	JUMP_CYCLE: [
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,3,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,4,3,3,3,0,0,0,0,0,
+			    0,0,0,0,0,3,3,3,3,3,3,3,0,0,0,0,
+			    0,0,0,0,0,4,4,4,4,2,3,2,0,0,0,0,
+			    0,0,0,0,4,4,4,4,2,4,4,1,0,0,0,0,
+			    0,0,0,0,4,4,4,2,2,4,4,0,0,0,0,0,
+			    0,0,0,0,4,4,4,4,0,0,3,4,0,0,0,0,
+			    0,0,0,0,4,4,4,2,4,2,4,3,0,0,0,0,
+			    0,0,0,0,0,4,4,4,4,4,0,0,0,0,0,0,
+			    0,0,0,0,0,0,2,3,3,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,0,3,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			frameDuration: METROID_JUMP_FRAMES[0],
+			tPad: 2,
+			bPad: 3,
+			lPad: 4,
+			rPad: 4,
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,4,4,0,0,0,0,0,0,
+			    0,0,0,0,0,4,4,4,4,4,4,3,0,0,0,0,
+			    0,0,0,3,2,4,4,4,4,4,4,3,3,0,0,0,
+			    0,0,0,0,3,4,2,4,2,4,4,3,4,3,0,0,
+			    0,0,0,3,3,4,4,0,2,2,4,3,3,3,0,0,
+			    0,0,0,3,0,4,2,0,4,4,2,3,3,3,0,0,
+			    0,0,0,0,0,0,4,3,4,4,3,3,3,0,0,0,
+			    0,0,0,0,0,0,3,4,0,1,2,3,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			frameDuration: METROID_JUMP_FRAMES[1],
+			tPad: 4,
+			bPad: 4,
+			lPad: 3,
+			rPad: 2,
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,3,0,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,3,3,2,0,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,4,4,4,0,0,0,0,0,
+			    0,0,0,0,3,4,2,4,2,4,4,4,0,0,0,0,
+			    0,0,0,0,4,3,0,0,4,4,4,4,0,0,0,0,
+			    0,0,0,0,0,4,4,2,2,4,4,4,0,0,0,0,
+			    0,0,0,0,1,4,4,2,4,4,4,4,0,0,0,0,
+			    0,0,0,0,2,3,2,4,4,4,4,0,0,0,0,0,
+			    0,0,0,0,3,3,3,3,3,3,3,0,0,0,0,0,
+			    0,0,0,0,0,3,3,3,4,3,0,0,0,0,0,0,
+			    0,0,0,0,0,0,3,3,3,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			frameDuration: METROID_JUMP_FRAMES[2],
+			tPad: 3,
+			bPad: 2,
+			lPad: 4,
+			rPad: 4,
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,3,2,1,0,4,3,0,0,0,0,0,0,
+			    0,0,0,3,3,3,4,4,3,4,0,0,0,0,0,0,
+			    0,0,3,3,3,2,4,4,0,2,4,0,3,0,0,0,
+			    0,0,3,3,3,4,2,2,0,4,4,3,3,0,0,0,
+			    0,0,3,4,3,4,4,2,4,2,4,3,0,0,0,0,
+			    0,0,0,3,3,4,4,4,4,4,4,2,3,0,0,0,
+			    0,0,0,0,3,4,4,4,4,4,4,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,4,4,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			frameDuration: METROID_JUMP_FRAMES[3],
+			tPad: 4,
+			bPad: 4,
+			lPad: 2,
+			rPad: 3,
+		},
+	],
+	TURN_CYCLE: [
+		{
+			sprite: [
+			    0,0,0,0,0,0,4,3,3,3,0,0,0,0,0,0,
+			    0,0,0,0,0,4,3,3,3,2,4,0,0,0,0,0,
+			    0,0,0,0,0,3,3,3,3,1,2,0,0,0,0,0,
+			    0,0,0,0,0,0,3,4,3,3,3,0,0,0,0,0,
+			    0,0,0,0,0,4,4,4,4,3,0,4,2,2,4,2,
+			    0,0,0,0,0,4,4,4,4,4,4,4,2,4,2,2,
+			    0,0,0,0,0,0,4,4,4,4,0,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,3,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,4,4,4,0,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,0,4,0,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,0,4,4,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,0,3,4,4,0,0,0,0,
+			    0,0,0,0,0,4,4,0,0,0,4,0,0,0,0,0,
+			    0,0,0,3,3,4,0,0,0,4,4,0,0,0,0,0,
+			    0,0,0,3,3,0,0,0,0,3,3,0,0,0,0,0,
+			    0,0,0,0,3,3,0,0,0,0,3,3,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 0,
+			bPad: 0,
+			lPad: 5,
+			rPad: 4,
+			frameDuration: 2,
+		},
+	],
+	CROUCH_CYCLE: [
+		{
+			sprite: [
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,4,3,3,3,0,0,0,0,0,0,
+			    0,0,0,0,0,4,3,3,3,2,4,0,0,0,0,0,
+			    0,0,0,0,0,3,3,3,3,1,2,0,0,0,0,0,
+			    0,0,0,0,0,0,3,4,3,3,3,0,2,2,4,2,
+			    0,0,0,0,0,4,4,4,4,3,0,4,2,4,2,2,
+			    0,0,0,0,3,4,4,4,4,4,4,2,0,0,0,0,
+			    0,0,0,0,2,0,3,3,2,4,4,0,0,0,0,0,
+			    0,0,0,0,0,4,4,3,3,0,4,0,0,0,0,0,
+			    0,0,0,3,3,4,0,0,0,4,4,0,0,0,0,0,
+			    0,0,0,3,3,0,0,0,0,3,3,0,0,0,0,0,
+			    0,0,0,0,3,3,0,0,0,0,3,3,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			tPad: 5,
+			bPad: 0,
+			lPad: 3,
+			rPad: 0,
+			frameDuration: 400,
+		},
+	],
+	SHOOT_CYCLE: [
+		{
+			sprite: [
+			    0,0,0,0,0,0,4,3,3,3,0,0,0,0,0,0,
+			    0,0,0,0,0,4,3,3,3,2,4,0,0,0,0,0,
+			    0,0,0,0,0,3,3,3,3,1,2,0,0,0,0,0,
+			    0,0,0,0,0,0,3,4,3,3,3,0,0,0,0,0,
+			    0,0,0,0,0,4,4,4,4,3,0,4,2,2,4,6,
+			    0,0,0,0,0,4,4,4,4,4,4,4,2,4,2,6,
+			    0,0,0,0,0,0,4,4,4,4,0,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,3,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,4,4,4,0,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,0,4,0,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,0,4,4,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,0,3,4,4,0,0,0,0,
+			    0,0,0,0,0,4,4,0,0,0,4,0,0,0,0,0,
+			    0,0,0,3,3,4,0,0,0,4,4,0,0,0,0,0,
+			    0,0,0,3,3,0,0,0,0,3,3,0,0,0,0,0,
+			    0,0,0,0,3,3,0,0,0,0,3,3,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			frameDuration: 3,
+			tPad: 0,
+			bPad: 0,
+			lPad: 3,
+			rPad: 0,
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,4,3,3,3,0,0,0,0,0,0,
+			    0,0,0,0,0,4,3,3,3,2,4,0,0,0,0,0,
+			    0,0,0,0,0,3,3,3,3,1,2,0,0,0,0,0,
+			    0,0,0,0,0,0,3,4,3,3,3,0,0,0,0,0,
+			    0,0,0,0,0,4,4,4,4,3,0,4,2,2,6,5,
+			    0,0,0,0,0,4,4,4,4,4,4,4,2,4,6,5,
+			    0,0,0,0,0,0,4,4,4,4,0,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,3,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,4,4,4,0,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,0,4,0,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,0,4,4,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,0,3,4,4,0,0,0,0,
+			    0,0,0,0,0,4,4,0,0,0,4,0,0,0,0,0,
+			    0,0,0,3,3,4,0,0,0,4,4,0,0,0,0,0,
+			    0,0,0,3,3,0,0,0,0,3,3,0,0,0,0,0,
+			    0,0,0,0,3,3,0,0,0,0,3,3,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			frameDuration: 3,
+			tPad: 0,
+			bPad: 0,
+			lPad: 3,
+			rPad: 0,
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,4,3,3,3,0,0,0,0,0,0,
+			    0,0,0,0,0,4,3,3,3,2,4,0,0,0,0,0,
+			    0,0,0,0,0,3,3,3,3,1,2,0,0,0,0,0,
+			    0,0,0,0,0,0,3,4,3,3,3,0,0,0,0,0,
+			    0,0,0,0,0,4,4,4,4,3,0,4,2,2,5,2,
+			    0,0,0,0,0,4,4,4,4,4,4,4,2,4,5,2,
+			    0,0,0,0,0,0,4,4,4,4,0,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,3,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,4,4,4,0,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,0,4,0,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,0,4,4,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,0,3,4,4,0,0,0,0,
+			    0,0,0,0,0,4,4,0,0,0,4,0,0,0,0,0,
+			    0,0,0,3,3,4,0,0,0,4,4,0,0,0,0,0,
+			    0,0,0,3,3,0,0,0,0,3,3,0,0,0,0,0,
+			    0,0,0,0,3,3,0,0,0,0,3,3,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			frameDuration: 5,
+			tPad: 0,
+			bPad: 0,
+			lPad: 3,
+			rPad: 0,
+		},
+		{
+			sprite: [
+			    0,0,0,0,0,0,4,3,3,3,0,0,0,0,0,0,
+			    0,0,0,0,0,4,3,3,3,2,4,0,0,0,0,0,
+			    0,0,0,0,0,3,3,3,3,1,2,0,0,0,0,0,
+			    0,0,0,0,0,0,3,4,3,3,3,0,0,0,0,0,
+			    0,0,0,0,0,4,4,4,4,3,0,4,2,2,4,2,
+			    0,0,0,0,0,4,4,4,4,4,4,4,2,4,2,2,
+			    0,0,0,0,0,0,4,4,4,4,0,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,3,0,0,0,0,0,0,0,
+			    0,0,0,0,0,0,0,4,4,4,0,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,0,4,0,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,0,4,4,0,0,0,0,0,
+			    0,0,0,0,0,0,4,4,0,3,4,4,0,0,0,0,
+			    0,0,0,0,0,4,4,0,0,0,4,0,0,0,0,0,
+			    0,0,0,3,3,4,0,0,0,4,4,0,0,0,0,0,
+			    0,0,0,3,3,0,0,0,0,3,3,0,0,0,0,0,
+			    0,0,0,0,3,3,0,0,0,0,3,3,0,0,0,0,
+			],
+			width: 16,
+			height: 16,
+			frameDuration: 3,
+			tPad: 0,
+			bPad: 0,
+			lPad: 3,
+			rPad: 0,
+		},
+	],
+}
+
+
+export const NEPTUNE_SPRITES = {
+	STAND_CYCLE: [
+		{
+			sprite: [
+				0, 0, 0, 0, 4, 4, 4, 4, 4, 4, 4, 0, 0, 0, 0, 0,
+				0, 0, 0, 4, 5, 5, 5, 5, 5, 5, 5, 4, 0, 0, 0, 0,
+				0, 0, 4, 5, 5, 5, 5, 5, 5, 5, 5, 4, 0, 0, 0, 0,
+				0, 0, 4, 5, 5, 4, 4, 4, 4, 4, 4, 4, 4, 0, 0, 0,
+				0, 0, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 0, 0,
+				0, 0, 4, 5, 4, 6, 6, 4, 6, 4, 6, 4, 0, 0, 0, 0,
+				0, 0, 4, 5, 4, 6, 5, 4, 6, 4, 5, 6, 4, 0, 0, 0,
+				0, 0, 0, 4, 5, 4, 6, 6, 6, 6, 6, 4, 4, 0, 0, 0,
+				0, 0, 0, 0, 4, 4, 4, 4, 4, 4, 4, 4, 0, 0, 0, 0,
+				0, 0, 0, 4, 4, 4, 5, 5, 5, 5, 4, 0, 0, 0, 0, 0,
+				0, 0, 4, 4, 4, 5, 5, 5, 5, 5, 5, 4, 4, 0, 0, 0,
+				0, 0, 4, 6, 6, 4, 5, 5, 5, 5, 5, 4, 6, 4, 0, 0,
+				0, 0, 4, 6, 6, 4, 4, 4, 4, 4, 4, 4, 6, 4, 0, 0,
+				0, 0, 0, 4, 4, 4, 4, 4, 0, 4, 4, 4, 4, 0, 0, 0,
+				0, 0, 4, 5, 5, 5, 4, 0, 0, 4, 5, 5, 5, 4, 0, 0,
+				0, 0, 4, 4, 4, 4, 4, 0, 0, 4, 4, 4, 4, 4, 0, 0,
+			],
+			width: 16,
+			height: 16,
+			frameDuration: 100,
+		},
+	],
+	TURN_CYCLE: [
+		{
+			sprite: [
+				0, 0, 0, 0, 0, 0, 4, 4, 4, 4, 4, 4, 4, 0, 0, 0,
+				0, 0, 0, 0, 4, 5, 5, 5, 5, 5, 5, 5, 4, 0, 0, 0,
+				0, 0, 0, 4, 5, 5, 5, 5, 5, 5, 5, 5, 4, 0, 0, 0,
+				0, 0, 0, 4, 5, 5, 4, 4, 4, 4, 4, 4, 4, 4, 0, 0,
+				0, 0, 0, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 0,
+				0, 0, 4, 4, 5, 4, 6, 6, 4, 6, 4, 6, 4, 0, 0, 0,
+				0, 4, 6, 4, 5, 4, 6, 5, 4, 6, 4, 5, 6, 4, 0, 0,
+				0, 4, 6, 6, 4, 5, 4, 6, 6, 6, 6, 6, 4, 4, 0, 0,
+				0, 0, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 0, 0, 0,
+				0, 0, 0, 0, 0, 4, 4, 6, 6, 4, 4, 4, 0, 0, 0, 0,
+				0, 0, 0, 0, 4, 4, 4, 6, 6, 4, 4, 4, 4, 0, 0, 0,
+				0, 0, 0, 4, 4, 5, 5, 4, 4, 4, 4, 4, 4, 0, 0, 0,
+				0, 0, 4, 5, 4, 5, 5, 4, 4, 4, 4, 4, 0, 0, 0, 0,
+				0, 0, 4, 5, 4, 4, 5, 5, 4, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 4, 5, 4, 4, 4, 0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 4, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+			],
+			width: 16,
+			height: 16,
+			frameDuration: 100,
+		},
+	],
+	JUMP_CYCLE: [
+		{
+			sprite: [
+				0, 0, 0, 0, 4, 4, 4, 4, 4, 4, 4, 0, 0, 0, 0, 0,
+				0, 0, 0, 4, 5, 5, 5, 5, 5, 5, 5, 4, 0, 0, 0, 0,
+				0, 0, 4, 5, 5, 5, 5, 4, 4, 4, 4, 4, 4, 0, 0, 0,
+				0, 0, 4, 5, 5, 4, 4, 4, 4, 4, 4, 4, 4, 4, 0, 0,
+				0, 0, 4, 5, 4, 4, 6, 6, 6, 6, 6, 4, 4, 0, 0, 0,
+				0, 0, 4, 4, 4, 6, 6, 4, 6, 4, 6, 4, 0, 0, 0, 0,
+				0, 0, 4, 4, 4, 6, 5, 4, 6, 4, 5, 6, 4, 4, 0, 0,
+				0, 4, 6, 6, 4, 4, 6, 6, 6, 6, 6, 4, 6, 6, 4, 0,
+				0, 4, 6, 6, 4, 4, 4, 4, 4, 4, 4, 4, 6, 6, 4, 0,
+				0, 4, 4, 4, 4, 4, 5, 5, 5, 5, 5, 4, 4, 4, 0, 0,
+				0, 0, 4, 4, 4, 4, 5, 5, 5, 4, 4, 4, 4, 0, 0, 0,
+				0, 0, 0, 0, 4, 4, 4, 5, 4, 4, 4, 5, 5, 4, 0, 0,
+				0, 0, 0, 4, 5, 4, 4, 4, 4, 4, 5, 5, 5, 4, 0, 0,
+				0, 0, 4, 5, 5, 4, 4, 4, 0, 4, 5, 5, 4, 0, 0, 0,
+				0, 0, 4, 5, 5, 4, 4, 0, 0, 0, 4, 4, 0, 0, 0, 0,
+				0, 0, 0, 4, 4, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+			],
+			width: 16,
+			height: 16,
+			frameDuration: undefined,
+		},
+	],
+	WALK_CYCLE: [
+		{
+			sprite: [
+				0, 0, 0, 0, 4, 4, 4, 4, 4, 4, 4, 0, 0, 0, 0, 0,
+				0, 0, 0, 4, 5, 5, 5, 5, 5, 5, 5, 4, 0, 0, 0, 0,
+				0, 0, 4, 5, 5, 5, 5, 5, 5, 5, 5, 4, 0, 0, 0, 0,
+				0, 0, 4, 5, 5, 4, 4, 4, 4, 4, 4, 4, 4, 0, 0, 0,
+				0, 0, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 0, 0,
+				0, 0, 4, 5, 4, 6, 6, 4, 6, 4, 6, 4, 0, 0, 0, 0,
+				0, 0, 4, 5, 4, 6, 5, 4, 6, 4, 5, 6, 4, 0, 0, 0,
+				0, 0, 0, 4, 5, 4, 6, 6, 6, 6, 6, 4, 4, 0, 0, 0,
+				0, 0, 0, 0, 4, 4, 4, 4, 4, 4, 4, 4, 0, 0, 0, 0,
+				0, 0, 0, 4, 4, 4, 4, 6, 6, 4, 4, 0, 0, 0, 0, 0,
+				0, 0, 0, 4, 4, 4, 4, 6, 6, 4, 4, 0, 0, 0, 0, 0,
+				0, 0, 0, 4, 4, 4, 4, 4, 4, 5, 4, 4, 0, 0, 0, 0,
+				0, 0, 0, 4, 4, 4, 5, 5, 5, 4, 4, 4, 4, 0, 0, 0,
+				0, 0, 4, 5, 4, 4, 4, 4, 4, 4, 5, 5, 4, 0, 0, 0,
+				0, 0, 4, 4, 5, 5, 4, 0, 4, 5, 5, 4, 4, 0, 0, 0,
+				0, 0, 0, 4, 4, 4, 4, 0, 0, 4, 4, 4, 0, 0, 0, 0,
+			],
+			width: 16,
+			height: 16,
+			frameDuration: 10,
+		},
+		{
+			sprite: [
+				0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 4, 4, 4, 4, 4, 4, 4, 0, 0, 0, 0, 0,
+				0, 0, 0, 4, 5, 5, 5, 5, 5, 5, 5, 4, 0, 0, 0, 0,
+				0, 0, 4, 5, 5, 5, 5, 5, 5, 5, 5, 4, 0, 0, 0, 0,
+				0, 0, 4, 5, 5, 4, 4, 4, 4, 4, 4, 4, 4, 0, 0, 0,
+				0, 0, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 0, 0,
+				0, 0, 4, 5, 4, 6, 6, 4, 6, 4, 6, 4, 0, 0, 0, 0,
+				0, 0, 4, 5, 4, 6, 5, 4, 6, 4, 5, 6, 4, 0, 0, 0,
+				0, 0, 0, 4, 5, 4, 6, 6, 6, 6, 6, 4, 4, 0, 0, 0,
+				0, 0, 0, 0, 4, 4, 4, 4, 4, 4, 4, 0, 0, 0, 0, 0,
+				0, 0, 0, 4, 4, 6, 6, 4, 5, 5, 4, 4, 0, 0, 0, 0,
+				0, 0, 0, 4, 4, 6, 6, 4, 5, 5, 5, 4, 4, 0, 0, 0,
+				0, 0, 0, 0, 4, 4, 4, 5, 5, 5, 4, 4, 4, 0, 0, 0,
+				0, 0, 0, 4, 5, 5, 4, 4, 4, 4, 4, 4, 0, 0, 0, 0,
+				0, 0, 0, 4, 4, 4, 4, 4, 5, 5, 5, 4, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 0, 0, 4, 4, 4, 4, 4, 0, 0, 0, 0,
+			],
+			width: 16,
+			height: 16,
+			frameDuration: 3,
+		},
+		{
+			sprite: [
+				0, 0, 0, 0, 4, 4, 4, 4, 4, 4, 4, 0, 0, 0, 0, 0,
+				0, 0, 0, 4, 5, 5, 5, 5, 5, 5, 5, 4, 0, 0, 0, 0,
+				0, 0, 4, 5, 5, 5, 5, 5, 5, 5, 5, 4, 0, 0, 0, 0,
+				0, 0, 4, 5, 5, 4, 4, 4, 4, 4, 4, 4, 4, 0, 0, 0,
+				0, 0, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 0, 0,
+				0, 0, 4, 5, 4, 6, 6, 4, 6, 4, 6, 4, 0, 0, 0, 0,
+				0, 0, 4, 5, 4, 6, 5, 4, 6, 4, 5, 6, 4, 0, 0, 0,
+				0, 0, 0, 4, 5, 4, 6, 6, 6, 6, 6, 4, 4, 0, 0, 0,
+				0, 0, 0, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 0, 0, 0,
+				0, 0, 4, 4, 4, 4, 5, 5, 5, 5, 4, 6, 6, 4, 0, 0,
+				0, 0, 4, 6, 6, 4, 5, 5, 5, 5, 5, 4, 6, 4, 0, 0,
+				0, 0, 4, 6, 6, 4, 4, 4, 4, 4, 4, 4, 4, 0, 0, 0,
+				0, 0, 0, 4, 4, 4, 4, 5, 5, 5, 4, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 4, 4, 5, 5, 5, 4, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 4, 4, 4, 4, 4, 4, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 0, 4, 4, 4, 4, 0, 0, 0, 0, 0, 0, 0,
+			],
+			width: 16,
+			height: 16,
+			frameDuration: 10,
+		},
+		{
+			sprite: [
+				0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+				0, 0, 0, 0, 4, 4, 4, 4, 4, 4, 4, 0, 0, 0, 0, 0,
+				0, 0, 0, 4, 5, 5, 5, 5, 5, 5, 5, 4, 0, 0, 0, 0,
+				0, 0, 4, 5, 5, 5, 5, 5, 5, 5, 5, 4, 0, 0, 0, 0,
+				0, 0, 4, 5, 5, 4, 4, 4, 4, 4, 4, 4, 4, 0, 0, 0,
+				0, 0, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 0, 0,
+				0, 0, 4, 5, 4, 6, 6, 4, 6, 4, 6, 4, 0, 0, 0, 0,
+				0, 0, 4, 5, 4, 6, 5, 4, 6, 4, 5, 6, 4, 0, 0, 0,
+				0, 0, 0, 4, 5, 4, 6, 6, 6, 6, 6, 4, 4, 0, 0, 0,
+				0, 0, 0, 4, 4, 4, 4, 4, 4, 4, 4, 4, 0, 0, 0, 0,
+				0, 0, 4, 4, 4, 4, 5, 5, 5, 5, 4, 6, 4, 0, 0, 0,
+				0, 0, 4, 4, 6, 6, 4, 5, 5, 5, 5, 4, 4, 0, 0, 0,
+				0, 0, 4, 4, 6, 6, 4, 5, 5, 4, 4, 4, 0, 0, 0, 0,
+				0, 0, 0, 4, 4, 4, 4, 4, 4, 4, 5, 4, 0, 0, 0, 0,
+				0, 0, 0, 0, 4, 5, 5, 5, 5, 4, 4, 4, 0, 0, 0, 0,
+				0, 0, 0, 0, 4, 4, 4, 4, 4, 4, 0, 0, 0, 0, 0, 0,
+			],
+			width: 16,
+			height: 16,
+			frameDuration: 3,
+		},
+	],
+};
